+++ ruitk/core/ruitk.js
@@ -1,0 +1,3 @@
+import "./elements.js";
+
+export { Ruitk } from "./core.js"